[#210] moving the reducer out of utils
Creating a new src/reducers directory to host our reducers. Adding an
intermediate directory for the version of the Cr it is targeting.

The payload objects were superfluous. Also the linter wasn't showing
properly the docstring of the subfields of the payload objects. Removing
them fixes this and creates a slightly better devX
