+++ src/configuration/broker-models.tsx
@@ -55,7 +55,7 @@ import {
   getCertManagerResourceTemplateFromAcceptor,
   getConfigSecret,
   listConfigs,
-} from '../brokers/utils';
+} from '../reducers/7.12/reducer';
 import { useTranslation } from '../i18n';
 import {
   CertIssuerModel,
